requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
python-multipart>=0.0.12
qdrant-client>=1.7.0
openai>=1.12.0
pypdf2>=3.0.0
python-dotenv>=1.0.0

